Move pathToInstalledPackage tests to fetch-unit.spec
